pyproject.toml
[tool.poetry]
name = "kotoamatsukami"
version = "0.1.0"
description = "An AI backed virtual control system"
authors = ["João Pedro Jacques Hoss <joaopedro_jh@hotmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
wheel = "^0.37.1"
setuptools = "^60.5.0"
tensorflow-gpu = "2.7.0"
numpy = "1.21.5"
numba = "0.55.0rc1"
Pillow = "^9.0.0"
pandas = "1.3.5"
opencv-python = "^4.5.5"
mediapipe = "^0.8.9"

[tool.poetry.dev-dependencies]
ipykernel = "^6.6.1"
plotly = "^5.5.0"
nbformat = "^5.1.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
